Move RapidShareExtractor to lib
