+++ lib/rapidshareextractor/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id("lib-android")
+}
+
+dependencies {
+    implementation(project(":lib:playlistutils"))
+}
