+++ src/en/yflix/build.gradle
@@ -1,7 +1,7 @@
 ext {
     extName = 'YFlix'
     extClass = '.YFlix'
-    extVersionCode = 4
+    extVersionCode = 3
     isNsfw = true
 }
 
@@ -9,4 +9,5 @@ apply from: "$rootDir/common.gradle"
 
 dependencies {
     implementation(project(':lib:playlistutils'))
+    implementation(project(':lib:rapidshareextractor'))
 }
