+++ src/en/yflix/src/eu/kanade/tachiyomi/animeextension/en/yflix/YFlix.kt
@@ -3,6 +3,7 @@ package eu.kanade.tachiyomi.animeextension.en.yflix
 import android.app.Application
 import android.content.SharedPreferences
 import androidx.preference.PreferenceScreen
+import aniyomi.lib.rapidshareextractor.RapidShareExtractor
 import eu.kanade.tachiyomi.animesource.ConfigurableAnimeSource
 import eu.kanade.tachiyomi.animesource.model.AnimeFilterList
 import eu.kanade.tachiyomi.animesource.model.AnimesPage
